Cleanup of old stores
