+++ web-common/src/features/models/utils/get-table-references/index.ts
@@ -1,6 +1,3 @@
-import type { FileArtifactsData } from "@rilldata/web-common/features/entity-management/file-artifacts-store";
-import type { V1CatalogEntry } from "@rilldata/web-common/runtime-client";
-
 export interface Reference {
   reference: string;
   type: "from" | "join";
@@ -47,50 +44,4 @@ export function getEmbeddedReferences(sql: string): Array<Reference> {
   }
 
   return embeddedSources;
-}
-
-export function getTableName(
-  ref: Reference,
-  entities: Record<string, FileArtifactsData>,
-): string {
-  if (!ref.reference) return "";
-  const tableRef = ref.reference.substring(1, ref.reference.length - 1);
-  if (!tableRef.match(/\//) && !ProtocolMatcher.test(tableRef))
-    return ref.reference;
-  return entities[tableRef]?.name ?? "";
-}
-
-function sourcePathMatchesReference(source: V1CatalogEntry, table: Reference) {
-  return (
-    `"${source.path}"` === table.reference ||
-    `'${source.path}'` === table.reference
-  );
-}
-
-function sourceTableReferenceIsEmbedded(
-  table: Reference,
-  embeddedSources: V1CatalogEntry[],
-) {
-  // check to see if the quoted version of this reference is in the embedded sources
-  return embeddedSources?.some((source) =>
-    sourcePathMatchesReference(source, table),
-  );
-}
-
-/**
- * Returns the name of the table reference, or the cached name of the embedded source.
- */
-export function getMatchingCatalogReference(
-  table: Reference,
-  embeddedSources: V1CatalogEntry[],
-  existingEntities: Record<string, FileArtifactsData>,
-) {
-  // if this reference is embedded, return the cached name
-  if (sourceTableReferenceIsEmbedded(table, embeddedSources)) {
-    return embeddedSources?.find((source) =>
-      sourcePathMatchesReference(source, table),
-    ).name;
-  } else {
-    return getTableName(table, existingEntities);
-  }
 }
