+++ web-common/src/features/models/workspace/inspector/ModelInspectorHeader.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import { getFilePathFromNameAndType } from "@rilldata/web-common/features/entity-management/entity-mappers";
-  import { newFileArtifactStore } from "@rilldata/web-common/features/entity-management/file-artifacts-store-new";
+  import { fileArtifactsStore } from "@rilldata/web-common/features/entity-management/file-artifacts-store";
   import { EntityType } from "@rilldata/web-common/features/entity-management/types";
   import {
     useModel,
@@ -34,7 +34,7 @@
   $: model = $modelQuery?.data?.model;
 
   $: modelPath = getFilePathFromNameAndType(modelName, EntityType.Model);
-  $: modelHasError = newFileArtifactStore.getFileHasErrors(
+  $: modelHasError = fileArtifactsStore.getFileHasErrors(
     queryClient,
     $runtime.instanceId,
     modelPath,
